requirements.txt
cleaninty~=0.1.3
discord.py~=2.2
jishaku~=2.5
requests~=2.28
